refactor: make printer module private
